+++ app/src/main/java/com/friendroids/moneymana/ui/NavigationActivity.kt
@@ -4,5 +4,5 @@ import androidx.annotation.DrawableRes
 
 interface NavigationActivity {
     fun setImageResource(@DrawableRes resId: Int)
-    fun openCategoryFragment(idCategory: Int)
+    fun openCategoryFragment(idCategory: Long)
 }
